requirements.txt
fastapi
uvicorn
requests
python-dotenv
SQLAlchemy
psycopg2-binary
httpx
websockets
